Initial groundwork for gallery-dl integration
