+++ core/src/main/java/module-info.java
@@ -40,6 +40,9 @@ module net.brlns.gdownloader {
 
     exports net.brlns.gdownloader;
     exports net.brlns.gdownloader.clipboard;
+    exports net.brlns.gdownloader.downloader;
+    exports net.brlns.gdownloader.downloader.enums;
+    exports net.brlns.gdownloader.downloader.structs;
     exports net.brlns.gdownloader.settings;
     exports net.brlns.gdownloader.settings.enums;
     exports net.brlns.gdownloader.settings.filters;
