+++ core/src/main/java/net/brlns/gdownloader/GDownloader.java
@@ -47,6 +47,8 @@ import javax.swing.plaf.FontUIResource;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 import net.brlns.gdownloader.clipboard.ClipboardManager;
+import net.brlns.gdownloader.downloader.DownloadManager;
+import net.brlns.gdownloader.lang.Language;
 import net.brlns.gdownloader.settings.Settings;
 import net.brlns.gdownloader.settings.enums.BrowserEnum;
 import net.brlns.gdownloader.ui.GUIManager;
@@ -65,8 +67,51 @@ import net.brlns.gdownloader.util.PriorityThreadPoolExecutor;
 import org.slf4j.helpers.FormattingTuple;
 import org.slf4j.helpers.MessageFormatter;
 
-import static net.brlns.gdownloader.Language.*;
+import static net.brlns.gdownloader.lang.Language.*;
 
+// TODO media converter
+// TODO implement CD Ripper
+// TODO d&d files for conversion to different formats, we already have ffmpeg anyway
+//
+// TODO max simultaneous downloads should be independent per website
+// TODO silence unnecessary debug messages
+// TODO investigate adding AppImage build
+// TODO scale on resolution DPI
+// TODO save last window size in config
+// TODO keep older versions of ytdlp and retry failed downloads against them
+// TODO Advanced users can edit the config.json directly to add extra yt-dlp arguments like proxy settings. but maybe expose those settings to the ui.
+// TODO verify checksums during updates, add bouncycastle, check signatures
+// TODO write a component factory for GUIManager
+// TODO git actions build for different platforms
+// FEEDBACK Should choose to download video and audio independently on each card
+// TODO maybe add notifications for each toggled toolbar option
+// DROPPED check updates on a timer, but do not ever restart when anything is in the queue.
+// TODO individual 'retry failed download' button
+// TODO --no-playlist when single video option is active
+// TODO Artifacting seems to be happening on the scroll pane with AMD video cards
+// TODO open a window asking which videos in a playlist to download or not
+// TODO RearrangeableDeque's offerLast should be linked to the cards in the UI
+// TODO Better visual eye candy for when dragging cards
+// TODO Add setting to allow the user to manually specify the target codec for audio transcoding? currently it defaults to aac.
+// TODO Add 'Clear Completed Downloads' button.
+// TODO Javadoc, a whole lot of it.
+// TODO Refactor this very class. Separate some logic into different methods.
+// TODO Twitch settings purposefully default to suboptimal quality due to huge file sizes. Maybe consider adding a warning about this in the GUI.
+// TODO Split GUI into a different subproject from core logic.
+// TODO Investigate screen reader support (https:// www.nvaccess.org/download/)
+// TODO Send notifications when a NO_METHOD is triggered, explaining why it was triggered.
+// TODO Test downloading sections of a livestream (currently it gets stuck on status PREPARING). Note: it also leaves a zombie ffmpeg process behind dealing with the hls stream.
+// TODO The issue above is a yt-dlp bug https:// github.com/yt-dlp/yt-dlp/issues/7927
+// TODO Add a viewer for log files.
+// TODO Implement rate-limiting options internally; the way it's currently implemented does not account for concurrent or non-playlist downloads.
+// TODO Notify the user whenever a setting that requires restart was changed.
+// TODO Quit lingering ffmpeg processes spawned by yt-dlp
+// TODO Verify which exceptions are important to display to the user via GDownloader::handleException
+// TODO Add an url ignore list / Allow filters to be disabled
+// TODO Add option to clear all installed updates and start fresh. (Tackling certain issues where failed updates could break downloads)
+// TODO Optional pluggable gallery-dl and curl integration
+// TODO Split main window from GUIManager
+// Off to a bootcamp, project on pause
 /**
  * GDownloader - GUI wrapper for yt-dlp
  *
@@ -104,7 +149,7 @@ public final class GDownloader {
     private ClipboardManager clipboardManager;
 
     @Getter
-    private YtDlpDownloader downloadManager;
+    private DownloadManager downloadManager;
 
     @Getter
     private List<AbstractGitUpdater> updaters = new ArrayList<>();
@@ -187,7 +232,7 @@ public final class GDownloader {
             clipboardManager = new ClipboardManager(this);
             clipboardManager.init();
 
-            downloadManager = new YtDlpDownloader(this);
+            downloadManager = new DownloadManager(this);
 
             guiManager = new GUIManager(this);
 
@@ -361,7 +406,7 @@ public final class GDownloader {
                 }
             }
 
-            if (downloadManager.getYtDlpPath() == null || !downloadManager.getYtDlpPath().exists()) {
+            if (!downloadManager.isMainDownloaderInitialized()) {
                 log.error("Failed to initialize YT-DLP, the program cannot continue. Exitting...");
 
                 if (!userInitiated) {
