+++ core/src/main/java/net/brlns/gdownloader/settings/enums/ISettingsEnum.java
@@ -18,7 +18,7 @@ package net.brlns.gdownloader.settings.enums;
 
 import java.util.Arrays;
 
-import static net.brlns.gdownloader.Language.*;
+import static net.brlns.gdownloader.lang.Language.*;
 
 /**
  * @author Gabriel / hstr0100 / vertx010
